New version of standard service tests, finally

+++ src/main/java/test/api/model/StravaActivityTest.java
@@ -28,7 +28,7 @@ public class StravaActivityTest extends BeanTest<StravaActivity> {
 		validateActivity(activity, activity.getId(), activity.getResourceState());
 	}
 
-	public static void validateActivity(final StravaActivity activity, final Integer id, final StravaResourceState state) {
+	public static void validateActivity(final StravaActivity activity, final Long id, final StravaResourceState state) {
 		assertNotNull(activity);
 		assertEquals(id, activity.getId());
 		assertEquals(state, activity.getResourceState());
@@ -43,7 +43,8 @@ public class StravaActivityTest extends BeanTest<StravaActivity> {
 			assertNotNull(activity.getAthlete());
 			assertTrue((activity.getAthlete().getResourceState() == StravaResourceState.META)
 					|| (activity.getAthlete().getResourceState() == StravaResourceState.SUMMARY));
-			StravaAthleteTest.validateAthlete(activity.getAthlete(), activity.getAthlete().getId(), activity.getAthlete().getResourceState());
+			StravaAthleteTest.validateAthlete(activity.getAthlete(), activity.getAthlete().getId(),
+					activity.getAthlete().getResourceState());
 			assertNotNull(activity.getName());
 			assertNotNull(activity.getDistance());
 			assertTrue(activity.getDistance() >= 0);
@@ -105,7 +106,8 @@ public class StravaActivityTest extends BeanTest<StravaActivity> {
 				StravaGearTest.validateGear(activity.getGear(), activity.getGearId(), activity.getGear().getResourceState());
 			}
 			assertNotNull(activity.getHasKudoed());
-			if ((activity.getType() == StravaActivityType.RIDE) && !activity.getManual() && !activity.getTrainer() && (activity.getCalories() != null)) {
+			if ((activity.getType() == StravaActivityType.RIDE) && !activity.getManual() && !activity.getTrainer()
+					&& (activity.getCalories() != null)) {
 				assertNotNull(activity.getKilojoules());
 			}
 			assertNotNull(activity.getKudosCount());
@@ -115,7 +117,8 @@ public class StravaActivityTest extends BeanTest<StravaActivity> {
 			assertNotNull(activity.getManual());
 			assertNotNull(activity.getMap());
 			if (!activity.getManual() && !activity.getTrainer()) {
-				StravaMapTest.validateMap(activity.getMap(), activity.getMap().getId(), activity.getMap().getResourceState(), activity);
+				StravaMapTest.validateMap(activity.getMap(), activity.getMap().getId(), activity.getMap().getResourceState(),
+						activity);
 			}
 			assertNotNull(activity.getMaxSpeed());
 			assertTrue(activity.getMaxSpeed() >= 0);
@@ -155,7 +158,8 @@ public class StravaActivityTest extends BeanTest<StravaActivity> {
 			assertNotNull(activity.getAthlete());
 			assertTrue((activity.getAthlete().getResourceState() == StravaResourceState.META)
 					|| (activity.getAthlete().getResourceState() == StravaResourceState.SUMMARY));
-			StravaAthleteTest.validateAthlete(activity.getAthlete(), activity.getAthlete().getId(), activity.getAthlete().getResourceState());
+			StravaAthleteTest.validateAthlete(activity.getAthlete(), activity.getAthlete().getId(),
+					activity.getAthlete().getResourceState());
 			assertNotNull(activity.getName());
 			assertNotNull(activity.getDistance());
 			assertTrue(activity.getDistance() >= 0);
@@ -195,7 +199,8 @@ public class StravaActivityTest extends BeanTest<StravaActivity> {
 			assertNull(activity.getGear());
 			// Optional assertNotNull(activity.getGearId());
 			assertNotNull(activity.getHasKudoed());
-			if ((activity.getType() == StravaActivityType.RIDE) && !activity.getManual() && !activity.getTrainer() && (activity.getCalories() != null)) {
+			if ((activity.getType() == StravaActivityType.RIDE) && !activity.getManual() && !activity.getTrainer()
+					&& (activity.getCalories() != null)) {
 				assertNotNull(activity.getKilojoules());
 			}
 			assertNotNull(activity.getKudosCount());
@@ -205,7 +210,8 @@ public class StravaActivityTest extends BeanTest<StravaActivity> {
 			assertNotNull(activity.getManual());
 			assertNotNull(activity.getMap());
 			if (!activity.getManual() && !activity.getTrainer()) {
-				StravaMapTest.validateMap(activity.getMap(), activity.getMap().getId(), activity.getMap().getResourceState(), activity);
+				StravaMapTest.validateMap(activity.getMap(), activity.getMap().getId(), activity.getMap().getResourceState(),
+						activity);
 			}
 			assertNotNull(activity.getMaxSpeed());
 			assertTrue(activity.getMaxSpeed() >= 0);
@@ -262,9 +268,9 @@ public class StravaActivityTest extends BeanTest<StravaActivity> {
 			assertNull(activity.getHasKudoed());
 			assertNull(activity.getKilojoules());
 			assertNull(activity.getKudosCount());
-//			assertNull(activity.getLocationCity());
-//			assertNull(activity.getLocationCountry());
-//			assertNull(activity.getLocationState());
+			// assertNull(activity.getLocationCity());
+			// assertNull(activity.getLocationCountry());
+			// assertNull(activity.getLocationState());
 			assertNull(activity.getManual());
 			assertNull(activity.getMap());
 			assertNull(activity.getMaxHeartrate());
