+++ src/main/java/test/api/rest/activity/ListActivityCommentsTest.java
@@ -5,20 +5,21 @@ import test.api.model.StravaCommentTest;
 import test.api.rest.APIListTest;
 import test.utils.TestUtils;
 
-public class ListActivityCommentsTest extends APIListTest<StravaComment, Integer> {
+public class ListActivityCommentsTest extends APIListTest<StravaComment, Long> {
 	/**
 	 *
 	 */
 	public ListActivityCommentsTest() {
 		this.listCallback = (api, id) -> api.listActivityComments(id, null, null, null);
-		this.pagingCallback = paging -> api().listActivityComments(TestUtils.ACTIVITY_WITH_COMMENTS, null, paging.getPage(), paging.getPageSize());
+		this.pagingCallback = paging -> api().listActivityComments(TestUtils.ACTIVITY_WITH_COMMENTS, null, paging.getPage(),
+				paging.getPageSize());
 	}
 
 	/**
 	 * @see test.api.rest.APIListTest#invalidId()
 	 */
 	@Override
-	protected Integer invalidId() {
+	protected Long invalidId() {
 		return TestUtils.ACTIVITY_INVALID;
 	}
 
@@ -26,7 +27,7 @@ public class ListActivityCommentsTest extends APIListTest<StravaComment, Integer
 	 * @see test.api.rest.APIListTest#privateId()
 	 */
 	@Override
-	protected Integer privateId() {
+	protected Long privateId() {
 		return TestUtils.ACTIVITY_PRIVATE;
 	}
 
@@ -34,7 +35,7 @@ public class ListActivityCommentsTest extends APIListTest<StravaComment, Integer
 	 * @see test.api.rest.APIListTest#privateIdBelongsToOtherUser()
 	 */
 	@Override
-	protected Integer privateIdBelongsToOtherUser() {
+	protected Long privateIdBelongsToOtherUser() {
 		return TestUtils.ACTIVITY_PRIVATE_OTHER_USER;
 	}
 
@@ -57,7 +58,7 @@ public class ListActivityCommentsTest extends APIListTest<StravaComment, Integer
 	 * @see test.api.rest.APIListTest#validId()
 	 */
 	@Override
-	protected Integer validId() {
+	protected Long validId() {
 		return TestUtils.ACTIVITY_FOR_AUTHENTICATED_USER;
 	}
 
@@ -65,7 +66,7 @@ public class ListActivityCommentsTest extends APIListTest<StravaComment, Integer
 	 * @see test.api.rest.APIListTest#validIdBelongsToOtherUser()
 	 */
 	@Override
-	protected Integer validIdBelongsToOtherUser() {
+	protected Long validIdBelongsToOtherUser() {
 		return TestUtils.ACTIVITY_FOR_UNAUTHENTICATED_USER;
 	}
 
@@ -73,7 +74,7 @@ public class ListActivityCommentsTest extends APIListTest<StravaComment, Integer
 	 * @see test.api.rest.APIListTest#validIdNoChildren()
 	 */
 	@Override
-	protected Integer validIdNoChildren() {
+	protected Long validIdNoChildren() {
 		return TestUtils.ACTIVITY_WITHOUT_COMMENTS;
 	}
