+++ src/main/java/test/api/rest/activity/ListActivityKudoersTest.java
@@ -7,7 +7,7 @@ import test.api.model.StravaAthleteTest;
 import test.api.rest.APIListTest;
 import test.utils.TestUtils;
 
-public class ListActivityKudoersTest extends APIListTest<StravaAthlete, Integer> {
+public class ListActivityKudoersTest extends APIListTest<StravaAthlete, Long> {
 	/**
 	 *
 	 */
@@ -20,7 +20,7 @@ public class ListActivityKudoersTest extends APIListTest<StravaAthlete, Integer>
 	 * @see test.api.rest.APIListTest#invalidId()
 	 */
 	@Override
-	protected Integer invalidId() {
+	protected Long invalidId() {
 		return TestUtils.ACTIVITY_INVALID;
 	}
 
@@ -28,7 +28,7 @@ public class ListActivityKudoersTest extends APIListTest<StravaAthlete, Integer>
 	 * @see test.api.rest.APIListTest#privateId()
 	 */
 	@Override
-	protected Integer privateId() {
+	protected Long privateId() {
 		return TestUtils.ACTIVITY_PRIVATE_WITH_KUDOS;
 	}
 
@@ -36,7 +36,7 @@ public class ListActivityKudoersTest extends APIListTest<StravaAthlete, Integer>
 	 * @see test.api.rest.APIListTest#privateIdBelongsToOtherUser()
 	 */
 	@Override
-	protected Integer privateIdBelongsToOtherUser() {
+	protected Long privateIdBelongsToOtherUser() {
 		return TestUtils.ACTIVITY_PRIVATE_OTHER_USER;
 	}
 
@@ -59,7 +59,7 @@ public class ListActivityKudoersTest extends APIListTest<StravaAthlete, Integer>
 	 * @see test.api.rest.APIListTest#validId()
 	 */
 	@Override
-	protected Integer validId() {
+	protected Long validId() {
 		return TestUtils.ACTIVITY_WITH_KUDOS;
 	}
 
@@ -67,7 +67,7 @@ public class ListActivityKudoersTest extends APIListTest<StravaAthlete, Integer>
 	 * @see test.api.rest.APIListTest#validIdBelongsToOtherUser()
 	 */
 	@Override
-	protected Integer validIdBelongsToOtherUser() {
+	protected Long validIdBelongsToOtherUser() {
 		return TestUtils.ACTIVITY_FOR_UNAUTHENTICATED_USER;
 	}
 
@@ -75,7 +75,7 @@ public class ListActivityKudoersTest extends APIListTest<StravaAthlete, Integer>
 	 * @see test.api.rest.APIListTest#validIdNoChildren()
 	 */
 	@Override
-	protected Integer validIdNoChildren() {
+	protected Long validIdNoChildren() {
 		return TestUtils.ACTIVITY_WITHOUT_KUDOS;
 	}
