+++ src/main/java/test/api/rest/activity/ListActivityPhotosTest.java
@@ -8,7 +8,7 @@ import test.api.rest.APIListTest;
 import test.issues.strava.Issue68;
 import test.utils.TestUtils;
 
-public class ListActivityPhotosTest extends APIListTest<StravaPhoto, Integer> {
+public class ListActivityPhotosTest extends APIListTest<StravaPhoto, Long> {
 	/**
 	 *
 	 */
@@ -24,7 +24,7 @@ public class ListActivityPhotosTest extends APIListTest<StravaPhoto, Integer> {
 	 * @see test.api.rest.APIListTest#invalidId()
 	 */
 	@Override
-	protected Integer invalidId() {
+	protected Long invalidId() {
 		return TestUtils.ACTIVITY_INVALID;
 	}
 
@@ -32,7 +32,7 @@ public class ListActivityPhotosTest extends APIListTest<StravaPhoto, Integer> {
 	 * @see test.api.rest.APIListTest#privateId()
 	 */
 	@Override
-	protected Integer privateId() {
+	protected Long privateId() {
 		return TestUtils.ACTIVITY_PRIVATE_WITH_PHOTOS;
 	}
 
@@ -40,7 +40,7 @@ public class ListActivityPhotosTest extends APIListTest<StravaPhoto, Integer> {
 	 * @see test.api.rest.APIListTest#privateIdBelongsToOtherUser()
 	 */
 	@Override
-	protected Integer privateIdBelongsToOtherUser() {
+	protected Long privateIdBelongsToOtherUser() {
 		return TestUtils.ACTIVITY_PRIVATE_OTHER_USER;
 	}
 
@@ -65,7 +65,7 @@ public class ListActivityPhotosTest extends APIListTest<StravaPhoto, Integer> {
 	 * @see test.api.rest.APIListTest#validId()
 	 */
 	@Override
-	protected Integer validId() {
+	protected Long validId() {
 		return TestUtils.ACTIVITY_WITH_PHOTOS;
 	}
 
@@ -73,7 +73,7 @@ public class ListActivityPhotosTest extends APIListTest<StravaPhoto, Integer> {
 	 * @see test.api.rest.APIListTest#validIdBelongsToOtherUser()
 	 */
 	@Override
-	protected Integer validIdBelongsToOtherUser() {
+	protected Long validIdBelongsToOtherUser() {
 		return TestUtils.ACTIVITY_FOR_UNAUTHENTICATED_USER;
 	}
 
@@ -81,7 +81,7 @@ public class ListActivityPhotosTest extends APIListTest<StravaPhoto, Integer> {
 	 * @see test.api.rest.APIListTest#validIdNoChildren()
 	 */
 	@Override
-	protected Integer validIdNoChildren() {
+	protected Long validIdNoChildren() {
 		return TestUtils.ACTIVITY_WITHOUT_PHOTOS;
 	}
